adding claim reward

+++ src/main/kotlin/rs/yettel/bms/models/Reward.kt
@@ -7,6 +7,6 @@ data class Reward(
     val id: Long? = null,
     val rewardName: String,
     val points: Int,
-    val eligibleUsers: List<Int> = emptyList(),
-    val usedByUsers: List<Int> = emptyList()
+    val eligibleUsers: List<Long> = emptyList(),
+    val usedByUsers: List<Long> = emptyList()
 )
